# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 森林防火无人机智能系统 API
  version: 1.0.0
  description: |
    统一 REST 接口规范（第一阶段：REST + MQTT + WebSocket，07号第3章）。

    **通用约定（07号第4~7章）**
    - 统一前缀 `/api/v1`；AI 服务接口前缀 `/ai/v1`（Backend↔AI Service 内部调用）。
    - 通用请求头：`Authorization: Bearer {token}`、`Content-Type: application/json`、
      `X-Request-Id: {uuid}`、`X-Client-Version: {version}`；内部服务调用增加
      `X-Service-Name` / `X-Service-Version`。
    - 统一响应包裹：`{code, message, data, requestId}`；`code=0` 成功，
      错误码清单见 `components.schemas.ApiCode`（40002=业务引用数据不存在，
      40401=REST路径资源不存在，判定规则"先判路径、再判业务引用"）。

    **WebSocket 实时订阅（07号第29章，非本 REST 规范范围）**
    - 端点：`/ws/uav`、`/ws/fire`、`/ws/mission`、`/ws/alert`。
    - 事件示例：fire.created / fire.confirmed / fire.updated / uav.online /
      uav.offline / uav.position.updated / mission.created / mission.started /
      mission.completed / mission.failed。

    **MQTT 统一主题（07号第24章，语义 owner=05号协议文档）**
    - `uav/{deviceId}/state` `/telemetry` `/media` `/event` `/mission`
      `/command` `/command/result`。
    - 所有消息必含 messageId / timestamp / deviceId / requestId，服务端幂等处理。

    **核心闭环接口链（07号第34章，第一阶段实现核心）**
    UAV_MEDIA → POST /ai/v1/detection → FireDetection → POST /ai/v1/localization
    → FirePoint → FireIncident → POST /api/v1/missions → POST /api/v1/dispatch
    → POST /api/v1/uavs/{id}/commands → UAV_COMMAND → UAV_COMMAND_RESULT
    → UAV_MEDIA → Verification

    **API设计原则（07号第33章）**：前端不得直调厂商接口；AI不得直接控制无人机；
    AI返回算法结果而非业务事件；Backend负责业务状态机；Gateway负责设备通信；
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/risk/areas:
    get:
      tags: [Risk]
      summary: 风险区域列表/图层
      description: >-
        返回网格自动划分的风险区域（risk_area，1km 网格 5×5）。level 过滤 RiskLevel；
        bbox 为 [minLng,minLat,maxLng,maxLat] 视口过滤。数据由 POST /risk/assess 生成。
      parameters:
        - name: level
          in: query
          schema: { type: string, enum: [HIGH, MEDIUM, LOW] }
        - name: bbox
          in: query
          schema: { type: string, description: "minLng,minLat,maxLng,maxLat" }
      responses:
        '200':
          description: 风险区域列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RiskArea'
  /api/v1/risk/areas/{areaId}:
    get:
      tags: [Risk]
      summary: 风险区域详情（五因子分项）
      parameters:
        - name: areaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 区域详情 + risk_feature 五因子
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/RiskArea'
                          - type: object
                            properties:
                              factors:
                                type: array
                                items:
                                  $ref: '#/components/schemas/RiskFactor'
        '40400':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/predictions/latest:
    get:
      tags: [Risk]
      summary: 最新火势预测（F11）
      parameters:
        - name: incidentId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 最新预测（30/60min 两档；无活跃火情时 data 为空数组）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FirePrediction'
  /api/v1/predictions/run:
    post:
      tags: [Risk]
      summary: 手动重跑预测（F11）
      description: 对指定活跃火情重跑 Level 1 椭圆扩散预测并落新版本；缺省取最新 ACTIVE 火情。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                incidentId: { type: string, format: uuid }
      responses:
        '200':
          description: 新预测结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FirePrediction'
  /api/v1/patrol-suggestions:
    get:
      tags: [Risk]
      summary: 巡检建议列表（F08）
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [SUGGESTED, DISPATCHED, DISMISSED, EXPIRED] }
      responses:
        '200':
          description: 建议列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PatrolSuggestion'
  /api/v1/patrol-suggestions/generate:
    post:
      tags: [Risk]
      summary: 从风险区域生成巡检建议（F08）
      description: >-
        按 risk_score 降序取前 N（constants.yaml#patrol_suggestion#top_n_suggestions）个
        风险区域，蛇形覆盖航点生成 patrol_area 记录（status=SUGGESTED）。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                areaIds:
                  type: array
                  items: { type: string, format: uuid }
                  description: "指定区域；缺省=全部按分数排序取前N"
      responses:
        '200':
          description: 新生成的建议
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PatrolSuggestion'
  /api/v1/patrol-suggestions/{suggestionId}/dispatch:
    post:
      tags: [Risk]
      summary: 建议下发成 RISK_PATROL 任务
      description: >-
        经 F09-Basic 评分选机创建 mission（mission_type=RISK_PATROL，risk_area_id 回填），
        航点写 mission_waypoint；patrol_area.status→DISPATCHED，reason JSONB 记录 missionId。
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 创建的任务
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MissionInfo'
        '40900':
          $ref: '#/components/responses/ResourceConflict'
  /api/v1/patrol-suggestions/{suggestionId}/dismiss:
    post:
      tags: [Risk]
      summary: 驳回巡检建议
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 驳回后的建议（status=DISMISSED）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatrolSuggestion'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token（30min过渡基线）+ Refresh Token（7d），见07号第35章。"

  # -------------------------------------------------------------------
...
  responses:
...
    NotFound:
      description: "40401 资源不存在（REST路径无对应资源）"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ResourceConflict:
      description: "40902 资源冲突"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    Envelope:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ApiCode'
        message:
          type: string
          example: success
        requestId:
          type: string
          example: REQ-001
    ApiCode:
      type: integer
      description: |
        统一状态码（07号第7章）。0=成功；40001参数错误；40002数据不存在（业务引用）；
        40003状态非法；40101未认证；40301无权限；40401资源不存在（REST路径）；
        40901重复请求；40902资源冲突；42901限流；50001服务异常；50002 AI异常；
        50003 UAV通信异常；50004厂商接口异常。
      enum: [0, 40001, 40002, 40003, 40101, 40301, 40401, 40901, 40902, 42901, 50001, 50002, 50003, 50004]
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              nullable: true

    # ---------- Auth ----------
...
    MissionType:
      type: string
      description: "任务业务类型六项（enums.yaml#MissionType；owner=02号第19章，本文档只引用）"
      enum: [PATROL, RISK_PATROL, FIRE_VERIFICATION, FIRE_TRACKING, THERMAL_INSPECTION, SUPPORT]
...
    AltitudeMode:
      type: string
      description: "目标点高度参考系（基线第39章）：禁止无参考系的裸高度值，altitude必须与altitudeMode同时出现"
      enum: [ELLIPSOID, RELATIVE_TO_TAKEOFF, AGL]
    TargetPoint:
      type: object
      required: [latitude, longitude, altitude, altitudeMode]
      properties:
        latitude: { type: number, format: double, example: 35.1234 }
        longitude: { type: number, format: double, example: 139.1234 }
        altitude: { type: number, format: double, description: 目标高度值, example: 150 }
        altitudeMode:
          $ref: '#/components/schemas/AltitudeMode'
    MissionStatus:
      type: string
      description: "任务执行状态（enums.yaml#MissionStatus；owner=04号）"
      enum: [CREATED, PENDING, APPROVED, ASSIGNED, EXECUTING, COMPLETED, FAILED, CANCELLED]
...
    MissionInfo:
      type: object
      properties:
        missionId: { type: string, example: MISSION-001 }
        missionType:
          $ref: '#/components/schemas/MissionType'
        priority: { type: integer, example: 90 }
        incidentId: { type: string }
        target:
          $ref: '#/components/schemas/TargetPoint'
        status:
          $ref: '#/components/schemas/MissionStatus'

    # ---------- Dispatch ----------
...
    RiskLevel:
      type: string
      description: "风险等级（enums.yaml#RiskLevel enums-v1.4）；阈值见 constants.yaml#risk_assessment#level_thresholds"
      enum: [HIGH, MEDIUM, LOW]
    RiskArea:
      type: object
      properties:
        areaId: { type: string, format: uuid }
        areaCode: { type: string, example: "GRID-R3-C2" }
        areaName: { type: string, example: "网格 R3-C2" }
        geometry:
          type: object
          description: "网格多边形（MultiPolygon GeoJSON，PostGIS geometry(MultiPolygon,4326)）"
        riskScore: { type: number, format: double, example: 87.5 }
        riskLevel:
          $ref: '#/components/schemas/RiskLevel'
        evaluationTime: { type: string, format: date-time }
    RiskFactor:
      type: object
      description: "五因子分项（risk_feature 行）；权重见 constants.yaml#risk_assessment#weights"
      properties:
        featureType: { type: string, example: HISTORICAL, description: "HISTORICAL/WEATHER/VEGETATION/TERRAIN/HUMAN_ACTIVITY" }
        featureValue: { type: number, format: double, example: 92.0 }
        weight: { type: number, format: double, example: 0.30 }
        contribution: { type: number, format: double, example: 27.6 }
        source: { type: string, example: "ai-service mock provider" }
...
    FirePrediction:
      type: object
      description: "火势预测（fire_prediction 行，F11 Level 1 规则模型）"
      properties:
        predictionId: { type: string, format: uuid }
        incidentId: { type: string, format: uuid }
        baseTime: { type: string, format: date-time }
        forecastMinutes: { type: integer, example: 30, description: "30/60 两档（constants.yaml#fire_prediction#forecast_horizons_min）" }
        predictedGeometry:
          type: object
          description: "扩散椭圆多边形（MultiPolygon GeoJSON）"
        predictedAreaSquareMeter: { type: number, format: double, example: 48200 }
        confidence: { type: number, format: double, example: 0.78 }
        environmentalInput:
          type: object
          description: "气象 mock 快照（tempC/humidityPct/windSpeedMps/windDirectionDeg；与DJI遥测分离）"
        predictionResult:
          type: object
          description: "模型过程数据（火头方向/长短轴/蔓延速度等）"
        createdAt: { type: string, format: date-time }
    PatrolSuggestionStatus:
      type: string
      description: "巡检建议状态（enums.yaml#PatrolSuggestionStatus enums-v1.4；patrol_area.status）"
      enum: [SUGGESTED, DISPATCHED, DISMISSED, EXPIRED]
    PatrolSuggestion:
      type: object
      description: "巡检建议（patrol_area 行，F08 生成）"
      properties:
        suggestionId: { type: string, format: uuid }
        riskAreaId: { type: string, format: uuid }
        areaCode: { type: string, example: "GRID-R3-C2" }
        riskScore: { type: number, format: double, example: 87.5 }
        geometry:
          type: object
          description: "覆盖区域多边形（Polygon GeoJSON）"
        waypoints:
          type: array
          description: "蛇形覆盖航点（估时 9m/s、航高 100m）"
          items:
            type: object
            properties:
              sequenceNo: { type: integer }
              latitude: { type: number, format: double }
              longitude: { type: number, format: double }
              altitude: { type: number, format: double, example: 100 }
        priority: { type: integer, example: 90 }
        estimatedDurationMin: { type: number, format: double, example: 8.5 }
        status:
          $ref: '#/components/schemas/PatrolSuggestionStatus'
        missionId: { type: string, format: uuid, description: "DISPATCHED 时回填（存 reason JSONB）" }
        generatedAt: { type: string, format: date-time }

    # ---------- Command ----------
...